docai/ilios-DocAI/pyproject.toml
[tool.poetry]
name = "ilios-docai"
version = "0.1.0"
description = "Document AI PoC for the iliOS."
authors = ["Alexander Deineha <odeine@softserveinc.com>", "Mateusz Kaczorek <mkaczo@softserveinc.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.13"
black = "24.8.0"
isort = "5.13.2"
interrogate = "^1.5.0"
jupyter = "^1.0.0"
flakehell = "0.9.0"
flake8 = "3.9.0"
google-cloud-aiplatform = "1.67.1"
google-cloud-documentai = "2.32.0"
pretty = "^0.1"
pypdf = "4.3.1"
google-cloud-storage = "2.18.2"
google-cloud = "0.34.0"
pandas = "^2.2.1"
python-levenshtein = "0.21.1"
langchainhub = "0.1.15"
chromadb = "0.4.24"
gcsfs = "^2024.2.0"
seaborn = "^0.13.2"
scikit-learn = "^1.4.1.post1"
nltk = "^3.8.1"
rouge-score = "^0.1.2"
watchdog = "^4.0.0"
docx2pdf = "^0.1.8"
openpyxl = "^3.1.2"
tqdm = "^4.66.2"
google-cloud-discoveryengine = "0.11.14"
grpcio-tools = "^1.62.1"
h5py = "^3.10.0"
google-cloud-secret-manager = "2.20.2"
thefuzz = "^0.22.1"
fuzzyset = "^0.1.2"
langchain-chroma = "^0.1.0"
lark = "^1.1.9"
pypdf2 = "^3.0.1"
boto3 = "^1.34.117"
langchain-aws = "^0.1.6"
anthropic = "^0.28.0"
requests = "^2.32.3"
flask = "3.0"
click = "^8.1.7"
google-cloud-build = "3.24.2"
google-cloud-run = "0.10.8"
google-auth = "2.34.0"
google-auth-oauthlib = "1.2.1"
locust = "^2.29.1"
pymupdf = "^1.24.9"
mlflow = "^2.15.0"
langchain = "0.2.16"
langchain-google-genai = "1.0.10"
langchain-google-vertexai = "1.0.10"
langchain-community = "0.2.16"
streamlit = "1.38.0"
fastapi = "0.114.2"
sqlalchemy = "2.0.35"
psycopg2-binary = "2.9.9"
pyyaml = "^6.0.2"
types-pyyaml = "^6.0.12.20240917"
snakeviz = "^2.2.0"
plotly = "^5.24.1"
pgvector = "0.2.5"
numpy = "1.26.4"
asyncio = "3.4.3"
asyncpg = "0.29.0"
cloud-sql-python-connector = "1.12.1"
langchain-google-cloud-sql-pg = "0.11.0"
pg8000 = "1.31.2"
psycopg2 = "2.9.10"
python-tds = "1.15.0"
sqlalchemy-pytds = "1.0.2"
langchain-postgres = "0.0.12"
psycopg = "3.2.3"
pydantic = "2.9.2"
pyjwt = "2.10.0"
passlib = "1.7.4"
mypy = "1.11.2"
pytest = "7.2.0"


[tool.poetry.group.chatbot.dependencies]
transformers = "4.46.1"
torch = "2.5.1"

[build-system]
requires = ["poetry-core>=1.7.1"]
build-backend = "poetry.core.masonry.api"

[tool.flakehell]
extended_default_ignore = []
exclude = []
format = "colored"
max_line_length = 88
show_source = true

[tool.flakehell.plugins]
pycodestyle = ["+*", "-W503", "-W605", "-E203", "-W293", "-W291"]
pyflakes = ["+*"]

[tool.isort]
line_length = 88
combine_as_imports = true
balanced_wrapping = true
lines_after_imports = 2
include_trailing_comma = true
multi_line_output = 3

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.local
  | \.cache
  | \.stamps
  | \.venv
)/
'''
